Extract ColumnsHashing HashMethod
