+++ src/Interpreters/SetVariants.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <Interpreters/AggregationCommon.h>
-#include <Common/ColumnsHashing.h>
+#include <Common/ColumnsHashing/HashMethod.h>
 #include <Common/assert_cast.h>
 #include <Common/Arena.h>
 #include <Common/HashTable/HashSet.h>
